requirements.txt
pandas
scikit-surprise
torch
numpy
sklearn